backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-genai==1.10.0
google-cloud-firestore==2.20.1
google-auth==2.38.0
python-multipart==0.0.17
python-dotenv==1.0.1
pydantic>=2.0.0
